requirements.txt
milatools==0.0.18
numpy==1.25.2
scikit-learn==1.3.0
scikit-splearn==1.2.1
scipy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
wandb==0.15.8
virtualenv===20.24.3

